backend/requirements.txt
# Core Django packages
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
python-dotenv

# Database
psycopg2-binary

# Anthropic API
anthropic

# Optional: MongoDB support (uncomment when ready to switch)
# djongo
# pymongo

# Development tools
python-decouple
gunicorn

# Logging and monitoring
sentry-sdk